Move iterators to internal
