+++ internal/iterators/filter/filter.go
@@ -9,7 +9,7 @@ import (
 	"unicode"
 	"unicode/utf8"
 
-	"github.com/clipperhouse/uax29/iterators/util"
+	"github.com/clipperhouse/uax29/internal/iterators/util"
 	"golang.org/x/text/unicode/rangetable"
 )
