+++ internal/iterators/iter.go
