+++ internal/iterators/scanner.go
