+++ internal/iterators/segmenter.go
