+++ internal/iterators/string_segmenter.go
