+++ internal/iterators/util/util.go
