+++ phrases/segmenter_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 	"unicode/utf8"
 
-	"github.com/clipperhouse/uax29/iterators"
+	"github.com/clipperhouse/uax29/internal/iterators"
 	"github.com/clipperhouse/uax29/phrases"
 )
